Codebase/requirements.txt
numpy
pandas
matplotlib
PyYAML
scipy
statsmodels
jdatetime
openpyxl
